src/index.ts
class PGSRenderer {
  private videoElement: HTMLVideoElement;
  private subtitles: { timestamp: number; text: string }[] = [];

  constructor(videoElement: HTMLVideoElement) {
    this.videoElement = videoElement;
  }

  async loadSubtitles(subtitleFile: ArrayBuffer) {
    // Parse subtitles
    this.subtitles = this.parsePGS(subtitleFile);

    // Render subtitles
    this.renderSubtitles();
  }

  private parsePGS(subtitleFile: ArrayBuffer): { timestamp: number; text: string }[] {
    const view = new DataView(subtitleFile);
    const subtitles: { timestamp: number; text: string }[] = [];

    // Basic parsing logic (placeholder, needs proper PGS parsing logic)
    let i = 0;
    while (i < view.byteLength) {
      const timestamp = view.getUint32(i, true); // Placeholder: Read 4 bytes for timestamp
      const textLength = view.getUint8(i + 4); // Placeholder: Read 1 byte for text length
      let text = '';
      for (let j = 0; j < textLength; j++) {
        text += String.fromCharCode(view.getUint8(i + 5 + j)); // Placeholder: Read text characters
      }
      subtitles.push({ timestamp, text });
      i += 5 + textLength; // Move to the next subtitle entry
    }

    return subtitles;
  }

  private renderSubtitles() {
    let track = this.videoElement.textTracks[0];

    // Create a new track if one does not exist
    if (!track) {
      track = this.videoElement.addTextTrack('subtitles', 'English', 'en');
    }

    track.mode = 'showing'; // Make the subtitles visible

    this.subtitles.forEach(subtitle => {
      const cue = new VTTCue(subtitle.timestamp, subtitle.timestamp + 5, subtitle.text); // Placeholder: duration of 5 seconds
      track?.addCue(cue); // Ensure track is not null before adding cues
    });
  }
}

export { PGSRenderer };
